refactor: round and action config out of state
and declare actions as children of rounds

ref https://github.com/mlevison/high-performance-teams-game/issues/37
