+++ src/config/index.ts
@@ -1,0 +1,4 @@
+import { GameActionId as GameActionIdT } from './rounds';
+export type GameActionId = GameActionIdT;
+
+export { rounds } from './rounds';
