+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const TOTAL_ROUNDS = 6;
+export const EFFECT_HIDDEN = Symbol('EFFECT_HIDDEN');
