+++ src/state/effects/index.ts
@@ -7,6 +7,5 @@ export type Effect = EffectT;
 export type GameEffect = GameEffectT;
 export type VisibleEffect = VisibleEffectT;
 
-export { EFFECT_HIDDEN } from './types';
 export { gameEffectList } from './effects';
 export { isEffect, isVisibleEffect } from './helpers';
