+++ src/state/gameActions/findGameActionById.ts
@@ -1,5 +1,6 @@
+import { GameActionId } from '../../config';
 import { GameAction } from './types';
-import { GameActionId, gameActions } from './gameActions';
+import { gameActions } from './gameActions';
 
 export function findGameActionById(gameActionId: GameActionId): GameAction {
   return gameActions.find((gameAction) => gameActionId === gameAction.id)!;
