+++ src/state/gameActions/index.ts
@@ -1,4 +1,3 @@
-import { GameActionId as GameActionIdT } from './gameActions';
 import { GameAction as GameActionT } from './types';
 import { GameActionWithStatus as GameActionWithStatusT } from './getAvailableGameActions';
 export {
@@ -10,6 +9,5 @@ export { getEffect } from './getEffect';
 export { getCost } from './getCost';
 export { isGameActionWithIcon, isGameActionWithImage } from './types';
 
-export type GameActionId = GameActionIdT;
 export type GameAction = GameActionT;
 export type GameActionWithStatus = GameActionWithStatusT;
