--- src/state/roundDescriptions/index.ts
@@ -1,6 +1,0 @@
-import { roundDescriptions } from './roundDescriptions';
-export { getRoundDescriptionEffects } from './getRoundDescriptionEffects';
-
-const startCapacity = roundDescriptions['1']?.effect?.([])?.capacity || 0;
-
-export { startCapacity, roundDescriptions };
